third_party/WebKit/LayoutTests/editing/selection/addRange-failures-expected.txt

CONSOLE ERROR: line 15: The given range is null.
CONSOLE WARNING: line 21: 'Range.detach' is now a no-op, as per DOM (http://dom.spec.whatwg.org/#dom-range-detach).
CONSOLE ERROR: line 22: Discontiguous selection is not supported.
CONSOLE ERROR: line 31: The given range does not belong to the current selection's document.
CONSOLE ERROR: line 44: The given range and the current selection belong to two different document fragments.
Test error handling of Selection.addRange().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Running: testNull
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
Running: testDetachedRange
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
Running: testOtherDocument
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
Running: testOtherDocumentFragments
PASS selection.rangeCount is 1
PASS selection.getRangeAt(0).startContainer === document.body is true
PASS selection.getRangeAt(0).startOffset is 0
PASS selection.getRangeAt(0).endContainer === document.body is true
PASS selection.getRangeAt(0).endOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
